+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Provide argument index to custom validator (`arg.js`)
 * Change interface of custom reporter (`tester.js`)
 * Invert arguments `target` & `current` (`version.js`)
+* Fix `getScriptName` function (`path.js`)
 
 ## [0.1.1]
 * Throw an error if an empty `cmdline` argument is passed to `Process` constructor (`process.js`)